requirements.txt
black
pylint
pytest
pytest-cov
fastapi
uvicorn[standard]
tiktoken
python-dotenv
openai
